[refactor] deleting user with cascade type orphan removal without the check in service

+++ src/main/java/br/rafaeros/smp/modules/user/service/UserService.java
@@ -1,6 +1,5 @@
 package br.rafaeros.smp.modules.user.service;
 
-import java.util.List;
 import java.util.Objects;
 
 import org.springframework.data.domain.Page;
@@ -21,15 +20,12 @@ import br.rafaeros.smp.modules.user.controller.dto.UserResponseDTO;
 import br.rafaeros.smp.modules.user.model.User;
 import br.rafaeros.smp.modules.user.model.enums.Role;
 import br.rafaeros.smp.modules.user.repository.UserRepository;
-import br.rafaeros.smp.modules.userdevice.model.UserDevice;
-import br.rafaeros.smp.modules.userdevice.repository.UserDeviceRepository;
 import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
 public class UserService implements UserDetailsService {
     private final UserRepository userRepository;
-    private final UserDeviceRepository userDeviceRepository;
     private final PasswordEncoder passwordEncoder;
     private final String DEFAULT_PASSWORD = "mudar@123";
 
@@ -166,10 +162,8 @@ public class UserService implements UserDetailsService {
     public void delete(Long id) {
         User user = findByIdInternal(id);
 
-        List<UserDevice> linkedDevices = userDeviceRepository.findAllDevicesListByUserId(user.getId());
-
-        if (!linkedDevices.isEmpty()) {
-            userDeviceRepository.deleteAll(linkedDevices);
+        if (user == null) {
+            throw new ResourceNotFoundException("Usuário nao encontrado com o ID: " + id);
         }
 
         userRepository.delete(user);
